Moved converter files into a converter directory; likewise with unit tests
