+++ medacy/tests/tools/__init__.py
@@ -1,1 +1,1 @@
-from .con_test_data.con_test import con_text, source_text
+from medacy.tests.tools.converters.con_test_data.con_test import con_text, source_text
